Preserve separate navigator for Login
