+++ src/navigation/StackNavigators/NoBottomTabStackNavigator.js
@@ -3,14 +3,9 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import CameraContainer from "components/Camera/CameraContainer";
-import ForgotPassword from "components/LoginSignUp/ForgotPassword";
-import LicensePhotos from "components/LoginSignUp/LicensePhotos";
-import Login from "components/LoginSignUp/Login";
-import SignUp from "components/LoginSignUp/SignUp";
-import SignUpConfirmation from "components/LoginSignUp/SignUpConfirmation";
 import GroupPhotosContainer from "components/PhotoImporter/GroupPhotosContainer";
 import PhotoGallery from "components/PhotoImporter/PhotoGallery";
-import { CloseButton, Heading4 } from "components/SharedComponents";
+import { Heading4 } from "components/SharedComponents";
 import Mortal from "components/SharedComponents/Mortal";
 import PermissionGateContainer, {
   AUDIO_PERMISSIONS,
@@ -20,7 +15,6 @@ import SoundRecorder from "components/SoundRecorder/SoundRecorder";
 import { t } from "i18next";
 import {
   hideHeader,
-  hideHeaderLeft,
   showCustomHeader,
   showHeader
 } from "navigation/navigationOptions";
@@ -31,26 +25,9 @@ import SharedStackScreens from "./SharedStackScreens";
 
 const Stack = createNativeStackNavigator( );
 
-const LoginCloseButton = ( ) => (
-  <CloseButton
-    handleClose={navigation => navigation.getParent( )?.goBack( )}
-    buttonClassName="mr-[-15px]"
-  />
-);
-
 const soundRecorderTitle = ( ) => (
   <Heading4 className="text-white">{t( "RECORD-NEW-SOUND" )}</Heading4>
 );
-
-const LOGIN_SCREEN_OPTIONS = {
-  headerTitle: "",
-  headerTransparent: true,
-  headerTintColor: "white",
-  contentStyle: {
-    backgroundColor: "black"
-  },
-  headerRight: LoginCloseButton
-};
 
 const CAMERA_SCREEN_OPTIONS = {
   ...hideHeader,
@@ -144,34 +121,6 @@ const NoBottomTabStackNavigator = ( ): Node => (
       />
     </Stack.Group>
     {SharedStackScreens( )}
-    {/* Login Stack Group */}
-    <Stack.Group
-      screenOptions={LOGIN_SCREEN_OPTIONS}
-    >
-      <Stack.Screen
-        name="Login"
-        component={Login}
-        options={{
-          ...hideHeaderLeft
-        }}
-      />
-      <Stack.Screen
-        name="SignUp"
-        component={SignUp}
-      />
-      <Stack.Screen
-        name="ForgotPassword"
-        component={ForgotPassword}
-      />
-      <Stack.Screen
-        name="LicensePhotos"
-        component={LicensePhotos}
-      />
-      <Stack.Screen
-        name="SignUpConfirmation"
-        component={SignUpConfirmation}
-      />
-    </Stack.Group>
   </Stack.Navigator>
 );
